Move `Nav` to `~/components`
