+++ src/pages/ProjectPage/EditorNav.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
-import { NavContainer } from '~/shared/components/Layout/Nav'
-import { LogoLink, Navbar, NavbarItem } from '~/shared/components/Layout/Nav.styles'
+import { NavContainer } from '~/components/Nav'
+import { LogoLink, Navbar, NavbarItem } from '~/components/Nav/Nav.styles'
 import Logo from '~/shared/components/Logo'
 import Button from '~/shared/components/Button'
 import { REGULAR } from '~/shared/utils/styled'
